Split codegen/expr.rs into sub-modules
Move pipe/stdlib dispatch to codegen/pipes.rs (297 lines) and
parse/mock codegen to codegen/parse_mock.rs (373 lines), reducing
expr.rs from 1493 to 838 lines. Purely structural - no behavior changes.
