+++ src/codegen/expr.rs
@@ -4,22 +4,6 @@ const DEEP_EQUAL_FN: &str = "__floeEq";
 const THROW_NOT_IMPLEMENTED: &str = "(() => { throw new Error(\"not implemented\"); })()";
 const THROW_UNREACHABLE: &str = "(() => { throw new Error(\"unreachable\"); })()";
 
-/// Alternative codegen templates for Map operations on plain-object Records.
-/// TS Record<K, V> maps to Floe Map<K, V>, but at runtime it's a plain object,
-/// not a JS Map, so we use bracket access instead of Map API methods.
-fn record_map_template(method: &str) -> Option<&'static str> {
-    match method {
-        "get" => Some("$0[$1]"),
-        "has" => Some("($1 in $0)"),
-        "keys" => Some("Object.keys($0)"),
-        "values" => Some("Object.values($0)"),
-        "entries" => Some("Object.entries($0)"),
-        "size" => Some("Object.keys($0).length"),
-        "isEmpty" => Some("Object.keys($0).length === 0"),
-        _ => None,
-    }
-}
-
 impl Codegen {
     // ── Expressions ──────────────────────────────────────────────
 
@@ -463,7 +447,7 @@ impl Codegen {
     // ── Stdlib Helpers ─────────────────────────────────────────
 
     /// Emit each argument via a sub-codegen, propagating `needs_deep_equal`, and collect output strings.
-    fn emit_arg_strings(&mut self, args: &[Arg]) -> Vec<String> {
+    pub(super) fn emit_arg_strings(&mut self, args: &[Arg]) -> Vec<String> {
         let mut arg_strings = Vec::new();
         for arg in args {
             let mut sub = self.sub_codegen();
@@ -480,7 +464,7 @@ impl Codegen {
     }
 
     /// Emit a single expression via a sub-codegen, propagating `needs_deep_equal`.
-    fn emit_expr_string(&mut self, expr: &Expr) -> String {
+    pub(super) fn emit_expr_string(&mut self, expr: &Expr) -> String {
         let mut sub = self.sub_codegen();
         sub.emit_expr(expr);
         if sub.needs_deep_equal {
@@ -490,289 +474,15 @@ impl Codegen {
     }
 
     /// Check a stdlib template for deep-equal usage and expand it with the given arg strings.
-    fn apply_stdlib_template(&mut self, template: &str, arg_strings: &[String]) -> String {
+    pub(super) fn apply_stdlib_template(
+        &mut self,
+        template: &str,
+        arg_strings: &[String],
+    ) -> String {
         if template.contains(DEEP_EQUAL_FN) {
             self.needs_deep_equal = true;
         }
         expand_codegen_template(template, arg_strings)
-    }
-
-    // ── Pipe Lowering ────────────────────────────────────────────
-
-    /// Try to emit a stdlib call. Returns Some(output) if the callee is a stdlib function.
-    fn try_emit_stdlib_call(&mut self, callee: &Expr, args: &[Arg]) -> Option<String> {
-        if let ExprKind::Member { object, field } = &callee.kind
-            && let ExprKind::Identifier(module) = &object.kind
-            && let Some(stdlib_fn) = self.stdlib.lookup(module, field)
-        {
-            // For Map.get(obj, key) etc., check if the first arg is a Record-backed Map
-            let template = if module == "Map"
-                && let Some(first_arg) = args.first()
-                && let Arg::Positional(first_expr) = first_arg
-                && self.is_record_map(&first_expr.ty)
-            {
-                record_map_template(field).unwrap_or(stdlib_fn.codegen)
-            } else {
-                stdlib_fn.codegen
-            };
-            let arg_strings = self.emit_arg_strings(args);
-            Some(self.apply_stdlib_template(template, &arg_strings))
-        } else {
-            None
-        }
-    }
-
-    /// Check if an expression's type is a Map that came from a TS Record
-    /// (plain object at runtime, not a JS Map).
-    fn is_record_map(&self, ty: &crate::checker::Type) -> bool {
-        matches!(ty, crate::checker::Type::RecordMap { .. })
-    }
-
-    /// Try to emit a stdlib call in pipe context (piped value is first arg).
-    fn try_emit_stdlib_pipe(
-        &mut self,
-        left: &Expr,
-        callee: &Expr,
-        extra_args: &[Arg],
-    ) -> Option<String> {
-        if let ExprKind::Member { object, field } = &callee.kind
-            && let ExprKind::Identifier(module) = &object.kind
-            && let Some(stdlib_fn) = self.stdlib.lookup(module, field)
-        {
-            let template = if module == "Map" && self.is_record_map(&left.ty) {
-                record_map_template(field).unwrap_or(stdlib_fn.codegen)
-            } else {
-                stdlib_fn.codegen
-            };
-            let left_str = self.emit_expr_string(left);
-            let mut arg_strings = vec![left_str];
-            arg_strings.extend(self.emit_arg_strings(extra_args));
-            Some(self.apply_stdlib_template(template, &arg_strings))
-        } else {
-            None
-        }
-    }
-
-    /// Try to emit a qualified for-block call in pipe context.
-    /// `row |> AccentRow.toModel(args)` → `AccentRow__toModel(row, args)`
-    fn try_emit_for_block_pipe(
-        &mut self,
-        left: &Expr,
-        type_name: &str,
-        field: &str,
-        args: &[Arg],
-    ) -> bool {
-        if let Some(mangled) = self
-            .for_block_fns
-            .get(&(type_name.to_string(), field.to_string()))
-        {
-            let name = self
-                .import_aliases
-                .get(mangled)
-                .cloned()
-                .unwrap_or_else(|| mangled.clone());
-            self.push(&name);
-            self.push("(");
-            self.emit_expr(left);
-            if !args.is_empty() {
-                self.push(", ");
-                self.emit_args(args);
-            }
-            self.push(")");
-            true
-        } else {
-            false
-        }
-    }
-
-    /// Try to resolve a bare function name in pipe context via type-directed stdlib lookup.
-    /// Uses the checker's type map to determine which stdlib module to use.
-    /// e.g., `arr |> length` → left is Array → use Array.length template.
-    fn try_emit_bare_stdlib_pipe(
-        &mut self,
-        left: &Expr,
-        callee: &Expr,
-        extra_args: &[Arg],
-    ) -> Option<String> {
-        if let ExprKind::Identifier(name) = &callee.kind {
-            // Don't shadow locally defined functions, unless the name
-            // is also a stdlib function (stdlib takes priority in pipes)
-            if self.local_names.contains(name.as_str())
-                && self.stdlib.lookup_by_name(name).is_empty()
-            {
-                return None;
-            }
-
-            // Resolve stdlib module from the left-hand type.
-            // 1. Known type → type-directed (disambiguates Array.length vs String.length)
-            // 2. Unknown/Var type or no entry → name-based fallback
-            let stdlib_fn = match crate::type_layout::type_to_stdlib_module(&left.ty) {
-                Some(module) => self
-                    .stdlib
-                    .lookup(module, name)
-                    // Fallback: name might be in a different module (e.g. tap is in Pipe, not Array)
-                    .or_else(|| self.stdlib.lookup_by_name(name).into_iter().next()),
-                None => self.stdlib.lookup_by_name(name).into_iter().next(),
-            }?;
-
-            let template = stdlib_fn.codegen.to_string();
-            let left_str = self.emit_expr_string(left);
-            let mut arg_strings = vec![left_str];
-            arg_strings.extend(self.emit_arg_strings(extra_args));
-            return Some(self.apply_stdlib_template(&template, &arg_strings));
-        }
-        None
-    }
-
-    pub(super) fn emit_pipe(&mut self, left: &Expr, right: &Expr) {
-        match &right.kind {
-            // Stdlib pipe: `arr |> Array.sort` or `arr |> Array.map(fn)`
-            // Also handles type-directed resolution: `arr |> map(fn)` → stdlib lookup by name
-            ExprKind::Call { callee, args, .. } if !has_placeholder_arg(args) => {
-                if let Some(output) = self.try_emit_stdlib_pipe(left, callee, args) {
-                    self.push(&output);
-                    return;
-                }
-                // Type-directed resolution: bare function name → check stdlib
-                if let Some(output) = self.try_emit_bare_stdlib_pipe(left, callee, args) {
-                    self.push(&output);
-                    return;
-                }
-                // Qualified for-block: `row |> AccentRow.toModel(args)` → `AccentRow__toModel(row, args)`
-                if let ExprKind::Member { object, field } = &callee.kind
-                    && let ExprKind::Identifier(type_name) = &object.kind
-                    && self.try_emit_for_block_pipe(left, type_name, field, args)
-                {
-                    return;
-                }
-                // Fall through to normal call handling below
-                let callee_alias = if let ExprKind::Identifier(name) = &callee.kind {
-                    self.import_aliases.get(name.as_str()).cloned()
-                } else {
-                    None
-                };
-                if let Some(alias) = callee_alias {
-                    self.push(&alias);
-                } else {
-                    self.emit_expr(callee);
-                }
-                self.push("(");
-                self.emit_expr(left);
-                if !args.is_empty() {
-                    self.push(", ");
-                    self.emit_args(args);
-                }
-                self.push(")");
-            }
-            ExprKind::Member { object, field } => {
-                // Bare stdlib: `arr |> Array.sort` (no args)
-                if let Some(output) = self.try_emit_stdlib_pipe(left, right, &[]) {
-                    self.push(&output);
-                    return;
-                }
-                // Qualified for-block: `row |> AccentRow.toModel` → `AccentRow__toModel(row)`
-                if let ExprKind::Identifier(type_name) = &object.kind
-                    && self.try_emit_for_block_pipe(left, type_name, field, &[])
-                {
-                    return;
-                }
-                // Fallback: treat as function call
-                self.emit_expr(right);
-                self.push("(");
-                self.emit_expr(left);
-                self.push(")");
-            }
-            // `a |> f(b, _, c)` → `f(b, a, c)` — placeholder replacement
-            ExprKind::Call { callee, args, .. } if has_placeholder_arg(args) => {
-                self.emit_expr(callee);
-                self.push("(");
-                for (i, arg) in args.iter().enumerate() {
-                    if i > 0 {
-                        self.push(", ");
-                    }
-                    match arg {
-                        Arg::Positional(expr) if matches!(expr.kind, ExprKind::Placeholder) => {
-                            self.emit_expr(left);
-                        }
-                        Arg::Positional(expr) => self.emit_expr(expr),
-                        Arg::Named { label, value } => {
-                            // Named args stay as-is in TS (but we erase labels in calls)
-                            if matches!(value.kind, ExprKind::Placeholder) {
-                                self.emit_expr(left);
-                            } else {
-                                let _ = label;
-                                self.emit_expr(value);
-                            }
-                        }
-                    }
-                }
-                self.push(")");
-            }
-            // `a |> parse<T>` — substitute piped value into parse
-            ExprKind::Parse { type_arg, value } if matches!(value.kind, ExprKind::Placeholder) => {
-                let substituted = Expr::synthetic(
-                    ExprKind::Parse {
-                        type_arg: type_arg.clone(),
-                        value: Box::new(left.clone()),
-                    },
-                    right.span,
-                );
-                self.emit_expr(&substituted);
-            }
-            // `a |> f` → `f(a)` — bare function (also check stdlib)
-            ExprKind::Identifier(name) => {
-                if let Some(output) = self.try_emit_bare_stdlib_pipe(left, right, &[]) {
-                    self.push(&output);
-                    return;
-                }
-                // Use aliased import name if available (avoids TDZ conflicts)
-                let alias = self.import_aliases.get(name.as_str()).cloned();
-                if let Some(alias) = alias {
-                    self.push(&alias);
-                } else {
-                    self.emit_expr(right);
-                }
-                self.push("(");
-                self.emit_expr(left);
-                self.push(")");
-            }
-            // Fallback: treat as function call
-            _ => {
-                self.emit_expr(right);
-                self.push("(");
-                self.emit_expr(left);
-                self.push(")");
-            }
-        }
-    }
-
-    // ── Partial Application ──────────────────────────────────────
-
-    fn emit_partial_application(&mut self, callee: &Expr, args: &[Arg]) {
-        // `add(10, _)` → `(_x) => add(10, _x)`
-        let param_name = "_x";
-        self.push(&format!("({param_name}) => "));
-        self.emit_expr(callee);
-        self.push("(");
-        for (i, arg) in args.iter().enumerate() {
-            if i > 0 {
-                self.push(", ");
-            }
-            match arg {
-                Arg::Positional(expr) if matches!(expr.kind, ExprKind::Placeholder) => {
-                    self.push(param_name);
-                }
-                Arg::Positional(expr) => self.emit_expr(expr),
-                Arg::Named { value, .. } => {
-                    if matches!(value.kind, ExprKind::Placeholder) {
-                        self.push(param_name);
-                    } else {
-                        self.emit_expr(value);
-                    }
-                }
-            }
-        }
-        self.push(")");
     }
 
     // ── Constructor → Object Literal ─────────────────────────────
@@ -846,374 +556,9 @@ impl Codegen {
         }
     }
 
-    // ── Parse<T> Validation Codegen ─────────────────────────────
-
-    fn emit_parse(&mut self, type_arg: &TypeExpr, value: &Expr) {
-        // Generate: (() => { const __v = <value>; <checks>; return { ok: true, value: __v as T }; })()
-        self.push("(() => { const __v = ");
-        self.emit_expr(value);
-        self.push("; ");
-        self.emit_parse_checks("__v", type_arg, "");
-        self.push(&format!(
-            "return {{ {OK_FIELD}: true as const, {VALUE_FIELD}: __v as "
-        ));
-        self.emit_type_expr(type_arg);
-        self.push(" }; })()");
-    }
-
-    /// Emit validation checks for a given accessor path against a type expression.
-    /// `accessor` is the JS expression to check (e.g., "__v", "(__v as any).name").
-    /// `path` is a human-readable path for error messages (e.g., "", "field 'name'").
-    fn emit_parse_checks(&mut self, accessor: &str, type_expr: &TypeExpr, path: &str) {
-        match &type_expr.kind {
-            TypeExprKind::Named {
-                name, type_args, ..
-            } => {
-                match name.as_str() {
-                    "string" => {
-                        self.emit_typeof_check(accessor, "string", path);
-                    }
-                    "number" => {
-                        self.emit_typeof_check(accessor, "number", path);
-                    }
-                    "boolean" => {
-                        self.emit_typeof_check(accessor, "boolean", path);
-                    }
-                    "Array" => {
-                        // Array.isArray check + element validation
-                        let err_prefix = if path.is_empty() {
-                            String::new()
-                        } else {
-                            format!("{path}: ")
-                        };
-                        self.push(&format!(
-                            "if (!Array.isArray({accessor})) return {{ {OK_FIELD}: false as const, {ERROR_FIELD}: new Error(\"{err_prefix}expected array, got \" + typeof {accessor}) }}; "
-                        ));
-                        if let Some(elem_type) = type_args.first() {
-                            let idx_var = format!("__i{}", accessor.len());
-                            let elem_accessor = format!("{accessor}[{idx_var}]");
-                            let elem_path = if path.is_empty() {
-                                format!("element [\" + {idx_var} + \"]")
-                            } else {
-                                format!("{path} element [\" + {idx_var} + \"]")
-                            };
-                            self.push(&format!(
-                                "for (let {idx_var} = 0; {idx_var} < {accessor}.length; {idx_var}++) {{ "
-                            ));
-                            self.emit_parse_checks(&elem_accessor, elem_type, &elem_path);
-                            self.push("} ");
-                        }
-                    }
-                    "Option" => {
-                        // Allow undefined or validate inner type
-                        if let Some(inner_type) = type_args.first() {
-                            self.push(&format!("if ({accessor} !== undefined) {{ "));
-                            self.emit_parse_checks(accessor, inner_type, path);
-                            self.push("} ");
-                        }
-                    }
-                    _ => {
-                        // Named type — look up in expr_types to find if it's a known record.
-                        // For now, just check it's an object (non-null).
-                        let err_prefix = if path.is_empty() {
-                            String::new()
-                        } else {
-                            format!("{path}: ")
-                        };
-                        self.push(&format!(
-                            "if (typeof {accessor} !== \"object\" || {accessor} === null) return {{ {OK_FIELD}: false as const, {ERROR_FIELD}: new Error(\"{err_prefix}expected object, got \" + typeof {accessor}) }}; "
-                        ));
-                    }
-                }
-            }
-            TypeExprKind::Record(fields) => {
-                // Check it's an object
-                let err_prefix = if path.is_empty() {
-                    String::new()
-                } else {
-                    format!("{path}: ")
-                };
-                self.push(&format!(
-                    "if (typeof {accessor} !== \"object\" || {accessor} === null) return {{ {OK_FIELD}: false as const, {ERROR_FIELD}: new Error(\"{err_prefix}expected object, got \" + typeof {accessor}) }}; "
-                ));
-                // Check each field
-                for field in fields {
-                    let field_accessor = format!("({accessor} as any).{}", field.name);
-                    let field_path = if path.is_empty() {
-                        format!("field '{}'", field.name)
-                    } else {
-                        format!("{path}.{}", field.name)
-                    };
-                    self.emit_parse_checks(&field_accessor, &field.type_ann, &field_path);
-                }
-            }
-            TypeExprKind::Array(inner) => {
-                let err_prefix = if path.is_empty() {
-                    String::new()
-                } else {
-                    format!("{path}: ")
-                };
-                self.push(&format!(
-                    "if (!Array.isArray({accessor})) return {{ {OK_FIELD}: false as const, {ERROR_FIELD}: new Error(\"{err_prefix}expected array, got \" + typeof {accessor}) }}; "
-                ));
-                let idx_var = format!("__i{}", accessor.len());
-                let elem_accessor = format!("{accessor}[{idx_var}]");
-                let elem_path = if path.is_empty() {
-                    format!("element [\" + {idx_var} + \"]")
-                } else {
-                    format!("{path} element [\" + {idx_var} + \"]")
-                };
-                self.push(&format!(
-                    "for (let {idx_var} = 0; {idx_var} < {accessor}.length; {idx_var}++) {{ "
-                ));
-                self.emit_parse_checks(&elem_accessor, inner, &elem_path);
-                self.push("} ");
-            }
-            TypeExprKind::StringLiteral(_)
-            | TypeExprKind::Function { .. }
-            | TypeExprKind::Tuple(_)
-            | TypeExprKind::TypeOf(_)
-            | TypeExprKind::Intersection(_) => {
-                // Can't validate string literals, functions, tuples, typeof, or intersections at runtime — skip
-            }
-        }
-    }
-
-    fn emit_typeof_check(&mut self, accessor: &str, expected: &str, path: &str) {
-        let err_prefix = if path.is_empty() {
-            String::new()
-        } else {
-            format!("{path}: ")
-        };
-        self.push(&format!(
-            "if (typeof {accessor} !== \"{expected}\") return {{ {OK_FIELD}: false as const, {ERROR_FIELD}: new Error(\"{err_prefix}expected {expected}, got \" + typeof {accessor}) }}; "
-        ));
-    }
-
-    // ── Mock codegen ─────────────────────────────────────────────
-
-    /// Emit a mock value for the given type expression.
-    /// `counter` is used to generate unique sequential values.
-    /// `overrides` provides named field overrides from `mock<T>(field: value)`.
-    fn emit_mock(&mut self, type_arg: &TypeExpr, overrides: &[Arg], counter: &mut usize) {
-        self.emit_mock_for_type(type_arg, overrides, counter, "");
-    }
-
-    fn emit_mock_for_type(
-        &mut self,
-        type_expr: &TypeExpr,
-        overrides: &[Arg],
-        counter: &mut usize,
-        field_name: &str,
-    ) {
-        match &type_expr.kind {
-            TypeExprKind::Named {
-                name, type_args, ..
-            } => match name.as_str() {
-                "string" => {
-                    *counter += 1;
-                    let label = if field_name.is_empty() {
-                        "string"
-                    } else {
-                        field_name
-                    };
-                    self.push(&format!("\"mock-{label}-{}\"", counter));
-                }
-                "number" => {
-                    *counter += 1;
-                    self.push(&format!("{}", counter));
-                }
-                "boolean" => {
-                    // Alternate true/false based on counter
-                    self.push(if (*counter).is_multiple_of(2) {
-                        "true"
-                    } else {
-                        "false"
-                    });
-                    *counter += 1;
-                }
-                "Array" => {
-                    if let Some(elem_type) = type_args.first() {
-                        self.push("[");
-                        self.emit_mock_for_type(elem_type, &[], counter, field_name);
-                        self.push("]");
-                    } else {
-                        self.push("[]");
-                    }
-                }
-                "Option" => {
-                    // Option<T> → Some(mock<T>) — emit the inner value
-                    if let Some(inner_type) = type_args.first() {
-                        self.emit_mock_for_type(inner_type, &[], counter, field_name);
-                    } else {
-                        self.push("undefined");
-                    }
-                }
-                _ => {
-                    // Named user type — look up in type_defs
-                    if let Some(type_def) = self.type_defs.get(name).cloned() {
-                        self.emit_mock_for_typedef(&type_def, name, overrides, counter);
-                    } else {
-                        // Unknown type — emit empty object
-                        self.push("{}");
-                    }
-                }
-            },
-            TypeExprKind::Record(fields) => {
-                self.push("{ ");
-                for (i, field) in fields.iter().enumerate() {
-                    if i > 0 {
-                        self.push(", ");
-                    }
-                    // Check if this field has an override
-                    let has_override = overrides.iter().find(|arg| {
-                        if let Arg::Named { label, .. } = arg {
-                            label == &field.name
-                        } else {
-                            false
-                        }
-                    });
-                    self.push(&format!("{}: ", field.name));
-                    if let Some(Arg::Named { value, .. }) = has_override {
-                        self.emit_expr(value);
-                    } else {
-                        self.emit_mock_for_type(&field.type_ann, &[], counter, &field.name);
-                    }
-                }
-                self.push(" }");
-            }
-            TypeExprKind::Array(inner) => {
-                self.push("[");
-                self.emit_mock_for_type(inner, &[], counter, field_name);
-                self.push("]");
-            }
-            TypeExprKind::Tuple(types) => {
-                self.push("[");
-                for (i, ty) in types.iter().enumerate() {
-                    if i > 0 {
-                        self.push(", ");
-                    }
-                    self.emit_mock_for_type(ty, &[], counter, "");
-                }
-                self.push("]");
-            }
-            TypeExprKind::StringLiteral(value) => {
-                self.push(&format!("\"{value}\""));
-            }
-            TypeExprKind::Function { .. }
-            | TypeExprKind::TypeOf(_)
-            | TypeExprKind::Intersection(_) => {
-                self.push("(() => { throw new Error(\"mock function\"); })");
-            }
-        }
-    }
-
-    fn emit_mock_for_typedef(
-        &mut self,
-        type_def: &TypeDef,
-        type_name: &str,
-        overrides: &[Arg],
-        counter: &mut usize,
-    ) {
-        match type_def {
-            TypeDef::Record(entries) => {
-                self.push("{ ");
-                let mut first = true;
-                for entry in entries {
-                    match entry {
-                        RecordEntry::Field(field) => {
-                            if !first {
-                                self.push(", ");
-                            }
-                            first = false;
-                            let has_override = overrides.iter().find(|arg| {
-                                if let Arg::Named { label, .. } = arg {
-                                    label == &field.name
-                                } else {
-                                    false
-                                }
-                            });
-                            self.push(&format!("{}: ", field.name));
-                            if let Some(Arg::Named { value, .. }) = has_override {
-                                self.emit_expr(value);
-                            } else {
-                                self.emit_mock_for_type(&field.type_ann, &[], counter, &field.name);
-                            }
-                        }
-                        RecordEntry::Spread(spread) => {
-                            // Spread in mock: recursively mock the spread type
-                            if let Some(TypeDef::Record(spread_entries)) =
-                                self.type_defs.get(&spread.type_name).cloned()
-                            {
-                                for spread_entry in &spread_entries {
-                                    if let RecordEntry::Field(field) = spread_entry {
-                                        if !first {
-                                            self.push(", ");
-                                        }
-                                        first = false;
-                                        let has_override = overrides.iter().find(|arg| {
-                                            if let Arg::Named { label, .. } = arg {
-                                                label == &field.name
-                                            } else {
-                                                false
-                                            }
-                                        });
-                                        self.push(&format!("{}: ", field.name));
-                                        if let Some(Arg::Named { value, .. }) = has_override {
-                                            self.emit_expr(value);
-                                        } else {
-                                            self.emit_mock_for_type(
-                                                &field.type_ann,
-                                                &[],
-                                                counter,
-                                                &field.name,
-                                            );
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-                self.push(" }");
-            }
-            TypeDef::Union(variants) => {
-                // Pick first variant
-                if let Some(variant) = variants.first() {
-                    if variant.fields.is_empty() {
-                        // Unit variant
-                        self.push(&format!("{{ {TAG_FIELD}: \"{}\" as const }}", variant.name));
-                    } else {
-                        self.push(&format!("{{ {TAG_FIELD}: \"{}\" as const", variant.name));
-                        for field in &variant.fields {
-                            let fname = field.name.clone().unwrap_or_else(|| "value".to_string());
-                            self.push(&format!(", {fname}: "));
-                            self.emit_mock_for_type(&field.type_ann, &[], counter, &fname);
-                        }
-                        self.push(" }");
-                    }
-                } else {
-                    self.push("{}");
-                }
-            }
-            TypeDef::StringLiteralUnion(variants) => {
-                // Pick first variant
-                if let Some(first) = variants.first() {
-                    self.push(&format!("\"{first}\""));
-                } else {
-                    self.push("\"\"");
-                }
-            }
-            TypeDef::Alias(type_expr) => {
-                // Newtype: mock the inner type
-                self.emit_mock_for_type(type_expr, overrides, counter, type_name);
-            }
-        }
-    }
-
     // ── Arguments (labels erased) ────────────────────────────────
 
-    fn emit_args(&mut self, args: &[Arg]) {
+    pub(super) fn emit_args(&mut self, args: &[Arg]) {
         for (i, arg) in args.iter().enumerate() {
             if i > 0 {
                 self.push(", ");
@@ -1462,7 +807,7 @@ impl Codegen {
 }
 
 /// Check if an expression tree contains an Await node.
-fn expr_contains_await(expr: &Expr) -> bool {
+pub(super) fn expr_contains_await(expr: &Expr) -> bool {
     match &expr.kind {
         ExprKind::Await(_) => true,
         ExprKind::Call { callee, args, .. } => {
